% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_cartes_obstacles.R
\name{generer_carte_obstacle}
\alias{generer_carte_obstacle}
\title{Générer des cartes de localisation des ouvrages ROE}
\usage{
generer_carte_obstacle(donnees_ouvrage, fichier, fond_carte = c("ign", "osm"))
}
\arguments{
\item{donnees_ouvrage}{données d'un ouvrage}

\item{flux_ign}{vecteur dont le premier élément correspond à la clé et le
deuxième au nom technique de la couche parmi \href{https://geoservices.ign.fr/services-web-experts}{les services WMS de l'IGN}}
}
\description{
Générer des cartes de localisation des ouvrages ROE
}
