% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_fiche_terrain.R
\name{generer_fiches_terrain}
\alias{generer_fiches_terrain}
\title{Générer des fiches terrain ROE}
\usage{
generer_fiches_terrain(
  donnees,
  codes_roe = NULL,
  date_export = Sys.Date(),
  dossier_sortie = getwd(),
  sous_dossiers = NULL,
  fond_carte = c("ign", "osm")
)
}
\arguments{
\item{donnees}{données issus d'un export ROE}

\item{codes_roe}{vecteur de codes ROE des obstacles pour lesquels on veut
générer des fiches terrain.}

\item{date_export}{date à laquelle l'export du ROE utilisé a été réalisée}

\item{dossier_sortie}{chemin du dossier dans lequel seront enregistrées les
fiches terrain générées.}

\item{sous_dossiers}{vecteur contenant les noms de colonnes du tableau
\code{donnees} qui serviront à hiérarchiser les enregistrements des fiches (e.g.
par département en spécifiant "dept_nom")}
}
\description{
Cette fonction permet de générer des fiches terrain pré-remplies avec les
données issues d'un export du ROE. Ces fiches pourront alors servir lors de
visites sur site pour la mise en qualité des données du ROE.
}
