% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_traitements_qgis.R
\name{install_traitements_qgis}
\alias{install_traitements_qgis}
\title{Copie les scripts de traitements pour être accessible dans QGIS}
\usage{
install_traitements_qgis(dest_folder, overwrite = TRUE)
}
\arguments{
\item{dest_folder}{dossier d'où les scripts R seront accessibles pour QGIS.
Typiquement, il s'agit d'un dossier \code{processing/rscripts} dans le dossier
du profil utilisé.}
}
\description{
Copie les scripts de traitements pour être accessible dans QGIS
}
