% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renommer_couche.R
\name{renommer_couche}
\alias{renommer_couche}
\title{Renommer les champs d'une couche obstacles pour utilisation des fonctions du
package {RROE}}
\usage{
renommer_couche(couche, tableau_correspondances = NULL)
}
\arguments{
\item{couche}{Couche géographique contenant les informations sur les
obstacles à l'écoulement}

\item{tableau_correspondances}{Chemin vers le fichier csv contenant la correspondance
entre les noms de champ de la couche et les noms de champ utilisés dans le
script}
}
\description{
Renommer les champs d'une couche obstacles pour utilisation des fonctions du
package {RROE}
}
