% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer_donnees.R
\name{telecharger_roe_sandre}
\alias{telecharger_roe_sandre}
\title{Télécharger le ROE}
\usage{
telecharger_roe_sandre(code_dpt = NULL, nom_bassin = NULL, code_roe = NULL)
}
\arguments{
\item{code_dpt}{character Code INSEE du ou départements}

\item{nom_bassin}{character Nom du bassin hydrographique parmi:
SEINE-NORMANDIE, RHIN-MEUSE, RHONE-MEDITERRANEE, LOIRE-BRETAGNE,
ADOUR-GARONNE, ARTOIS-PICARDIE, CORSE}

\item{code_roe}{character Code du ou des ouvrages ROE}
}
\value{
a sf data frame
}
\description{
Cette fonction permet de télécharger le ROE diffusé via un \href{https://www.sandre.eaufrance.fr/atlas/srv/fre/catalog.search#/metadata/59057026-b40c-4cf9-9e3e-7296e0aa1a78}{fluw WFS du Sandre}.
}
\details{
La requête peut être filtrée pour un ou plusieurs départements, pour un ou
plusieurs bassins hydrographiques et pour un ou plusieurs ouvrages en
particulier. Si un code ROE est fourni, les autres filtres (département et
bassin) sont ignorés. Si des filtres sont fournis à la fois sur les
départements et les bassins sont fournis, le résultat correspond aux ouvrages
répondant à l'ensemble des filtres (opérateur ET).
}
